Conflicts:
	gem-light-dqm/dqm-reader/src/common/gem-read-events.cxx
	gemdaq-testing/gemreadout/include/gem/readout/GEMDataAMCformat.h
	gemdaq-testing/gemreadout/include/gem/readout/GEMDataParker.h
	gemdaq-testing/gemreadout/src/common/GEMDataParker.cc
	gemdaq-testing/gemsupervisor/src/common/GEMGLIBSupervisorWeb.cc

+++ gem-light-dqm/gemtreewriter/src/common/gemTreeComposer.cxx
@@ -222,7 +222,7 @@ int main(int argc, char** argv)
         uint32_t ZSFlag  = (0xffffff0000000000 & geb.header) >> 40; 
         uint16_t ChamID  = (0x000000fff0000000 & geb.header) >> 28; 
         uint32_t sumVFAT = (0x000000000fffffff & geb.header);
-        uint16_t  BX     = 0;
+        uint32_t BX      = 0;
 
         if (InpType == "Hex") {
           if(!gem::readout::GEMDataAMCformat::readGEBrunhed(inpf, geb)) break;
